backend/requirements.txt
sqlalchemy==1.3.17
pymysql==0.9.3
pyjwt==1.7.1
passlib[bcrypt]
fastapi[all]